fix: rename dao-types-v2 to dao-types
There is no v2 *hand wave*
