+++ src/stacks-contracts/deploy-dao.ts
@@ -19,7 +19,7 @@ import { DaoContractGenerator } from "./services/dao-contract-generator";
 import {
   ExpectedContractGeneratorArgs,
   getNetworkNameFromType,
-} from "./types/dao-types-v2";
+} from "./types/dao-types";
 import { DaoContractDeployer } from "./services/dao-contract-deployer";
 
 const usage = `Usage: bun run deploy-dao.ts <tokenSymbol> <tokenName> <tokenMaxSupply> <tokenUri> <logoUrl> <originAddress> <daoManifest> <tweetOrigin> <daoManifestInscriptionId> <generateFiles>`;
